+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - added direct support for Kimi Coding Plan provider
 - improved UX for initial user message
 - improved handling files within the context of the agent
+- properly removing the tool calls from assistant message when using remove messages up to here action
 
 ## [0.51.0]